pyproject.toml
[tool.poetry]
name = "insectid"
version = "0.1.0"
description = ""
authors = ["sebastian.cortes <sebitas0718@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flet = "^0.24.1"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
motor = "^3.6.0"
pydantic = "^2.9.2"
google-cloud-aiplatform = "^1.67.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
authlib = "^1.3.2"
itsdangerous = "^2.2.0"
google-cloud-firestore = "^2.19.0"
python-multipart = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
